More Documentation and Better Events
Improved documentation. Additionally, added the is_collided method to the Widget trait, allowing for an easy check of collision. This was also extended to a method in Scene called is_click_blocked, which checks all child nodes to determine if a click at a given position is blocked by a child/popup, making multi-tier UI layouts much easier to handle in terms of events.